Fixes #6708 - Use foreman org routes instead of adding new ones
